backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.16.3
google-generativeai==0.3.2
